+++ src/ebonite/ext/lightgbm/__init__.py
@@ -1,0 +1,2 @@
+from .dataset import LightGBMDatasetType, LightGBMDatasetHook
+from .model import LightGBMModelHook, LightGBMModelWrapper