+++ tests/ext/lightgbm/conftest.py
@@ -1,7 +1,7 @@
+import lightgbm as lgb
 import numpy as np
 import pandas as pd
 import pytest
-import lightgbm as lgb
 
 
 @pytest.fixture
@@ -11,7 +11,7 @@ def np_payload():
 
 @pytest.fixture
 def dataset_np(np_payload):
-    return lgb.Dataset(np_payload, label=np_payload)
+    return lgb.Dataset(np_payload, label=np_payload.reshape((-1,)), free_raw_data=False)
 
 
 @pytest.fixture
@@ -21,4 +21,4 @@ def df_payload():
 
 @pytest.fixture
 def dataset_df(df_payload):
-    return lgb.Dataset(df_payload, label=np.linspace(0, 2).reshape((-1, 1)))
+    return lgb.Dataset(df_payload, label=np.linspace(0, 2).reshape((-1, 1)), free_raw_data=False)