+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ Changelog
 Current release candidate
 -------------------------
 
+* Added support for LightGBM models
 * Added support for XGBoost models
 
 0.2.1 (2019-11-19)